Fix volume/structure desync and move scale_lattice_matrix to $lib/math
Gate supercell_volume on !supercell_loading so tiled volume and
supercell structure update in the same frame for large supercells
(which defer structure generation via setTimeout).

Move scale_lattice_matrix from structure/supercell to $lib/math where
it belongs as a general matrix operation, removing the isosurface →
structure cross-module dependency. Re-export from supercell for
backward compatibility.
